requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.5
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
jinja2==3.1.2
emails==0.6
asyncio==3.4.3
aiosmtplib==2.0.2
aioimaplib==1.0.1
pytest==7.4.3
httpx==0.25.1 
email_validator